mentcare_backend/requirements.txt
altgraph==0.17.5
asgiref==3.11.0
auto-py-to-exe==2.48.1
bottle==0.13.4
bottle-websocket==0.2.9
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clr_loader==0.2.10
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Eel==0.18.2
future==1.0.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.3.0
idna==3.11
importlib_resources==6.5.2
mysqlclient==2.2.7
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
proxy_tools==0.1.0
psycopg2-binary==2.9.11
pycparser==2.23
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.11.0
PyMySQL==1.1.2
pyparsing==3.2.5
pythonnet==3.0.5
pywebview==6.1
pywin32-ctypes==0.2.3
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
zope.event==6.1
zope.interface==8.1.1
